Centralize bitcoin-e2e test and refactor service creation
